Test main in conformance tests
This makes two changes to conformance testing.

Firstly we've added support to run a setup function after project generation
but before snapshots and testing. This allows us to write a test that sets up
"main" in a Pulumi.yaml file and test that "program directory" and "entry
point" are respected. Doing this correctly also required some bug fixes in the
conformance interface itself.

I've also added a `DisableSnapshotWriting` flag to the conformance interface so
that the tests we write for conformance itself that generate bad snapshots can
opt out of `PULUMI_ACCEPT` behaviour. I tripped myself up earlier by running
all the tests with `PULUMI_ACCEPT=1` and then getting confused on why the bad
snapshot tests were suddenly failing.

+++ cmd/pulumi-test-language/interface.go
@@ -99,6 +99,9 @@ type languageTestServer struct {
 	cancel chan bool
 	done   chan error
 	addr   string
+
+	// Used by _bad snapshot_ tests to disable snapshot writing.
+	DisableSnapshotWriting bool
 }
 
 func (eng *languageTestServer) Address() string {
@@ -323,8 +326,8 @@ func compareDirectories(actualDir, expectedDir string, allowNewFiles bool) ([]st
 
 // Do a snapshot check of the generated source code against the snapshot code. If PULUMI_ACCEPT is true just
 // write the new files instead.
-func doSnapshot(sourceDirectory, snapshotDirectory string) ([]string, error) {
-	if cmdutil.IsTruthy(os.Getenv("PULUMI_ACCEPT")) {
+func doSnapshot(disableSnapshotWriting bool, sourceDirectory, snapshotDirectory string) ([]string, error) {
+	if !disableSnapshotWriting && cmdutil.IsTruthy(os.Getenv("PULUMI_ACCEPT")) {
 		// Write files
 		err := os.RemoveAll(snapshotDirectory)
 		if err != nil {
@@ -697,7 +700,7 @@ func (eng *languageTestServer) RunLanguageTest(
 		}
 
 		snapshotDir := filepath.Join(token.SnapshotDirectory, "sdks", sdkName)
-		validations, err := doSnapshot(sdkTempDir, snapshotDir)
+		validations, err := doSnapshot(eng.DisableSnapshotWriting, sdkTempDir, snapshotDir)
 		if err != nil {
 			return nil, fmt.Errorf("sdk snapshot validation for %s: %w", pkg, err)
 		}
@@ -760,8 +763,17 @@ func (eng *languageTestServer) RunLanguageTest(
 		return makeTestResponse(fmt.Sprintf("generate project: %v", diagnostics)), nil
 	}
 
+	// We've generated the project, run the setup function to allow the test to modify the project before snapshotting
+	// and testing it.
+	if test.setup != nil {
+		err = test.setup(projectDir)
+		if err != nil {
+			return nil, fmt.Errorf("test setup: %w", err)
+		}
+	}
+
 	snapshotDir := filepath.Join(token.SnapshotDirectory, "projects", req.Test)
-	validations, err := doSnapshot(projectDir, snapshotDir)
+	validations, err := doSnapshot(eng.DisableSnapshotWriting, projectDir, snapshotDir)
 	if err != nil {
 		return nil, fmt.Errorf("program snapshot validation: %w", err)
 	}
@@ -769,24 +781,30 @@ func (eng *languageTestServer) RunLanguageTest(
 		return makeTestResponse(fmt.Sprintf("program snapshot validation failed:\n%s", strings.Join(validations, "\n"))), nil
 	}
 
-	// TODO(https://github.com/pulumi/pulumi/issues/13941): We don't capture stdout/stderr from the language
-	// plugin, so we can't show it back to the test.
+	project, err := workspace.LoadProject(filepath.Join(projectDir, "Pulumi.yaml"))
+	if err != nil {
+		return makeTestResponse(fmt.Sprintf("load project: %v", err)), nil
+	}
+
+	info := &engine.Projinfo{Proj: project, Root: projectDir}
+	pwd, main, err := info.GetPwdMain()
+	if err != nil {
+		return makeTestResponse(fmt.Sprintf("get pwd main: %v", err)), nil
+	}
+
 	programInfo := plugin.NewProgramInfo(
 		projectDir, /* rootDirectory */
-		projectDir, /* programDirectory */
-		"index",
+		pwd,        /* programDirectory */
+		main,
 		map[string]interface{}{})
 
+	// TODO(https://github.com/pulumi/pulumi/issues/13941): We don't capture stdout/stderr from the language
+	// plugin, so we can't show it back to the test.
 	err = languageClient.InstallDependencies(programInfo)
 	if err != nil {
 		return makeTestResponse(fmt.Sprintf("install dependencies: %v", err)), nil
 	}
 	// TODO(https://github.com/pulumi/pulumi/issues/13942): This should only add new things, don't modify
-
-	project, err := workspace.LoadProject(filepath.Join(projectDir, "Pulumi.yaml"))
-	if err != nil {
-		return makeTestResponse(fmt.Sprintf("load project: %v", err)), nil
-	}
 
 	// Create a temp dir for the a filestate backend to run in for the test
 	backendDir := filepath.Join(token.TemporaryDirectory, "backends", req.Test)
